Refactoring: 알람 구축 및 구축에 따른 리펙토링
