+++ backend/src/main/java/com/api/backend/notification/controller/NotificationController.java
@@ -1,12 +1,10 @@
 package com.api.backend.notification.controller;
 
-import com.api.backend.notification.data.type.NotificationUserType;
-import com.api.backend.notification.service.NotificationService;
+import com.api.backend.notification.service.EmitterService;
 import java.security.Principal;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
@@ -14,14 +12,14 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 @RequiredArgsConstructor
 public class NotificationController {
 
-  private final NotificationService notificationService;
+  private final EmitterService emitterService;
 
-  @GetMapping(value = "/notification/team/{teamId}", produces = "text/event-stream")
+  @GetMapping(value = "/subscribe/team/{teamId}", produces = "text/event-stream")
   public SseEmitter subscribeTeamRequest(
       Principal principal,
       @PathVariable(value = "teamId") Long teamId
   ) {
-    return notificationService.setTeamParticipantEmitter(
+    return emitterService.setTeamParticipantEmitter(
         teamId, Long.valueOf(principal.getName())
     );
   }
